+++ src/Discore/DiscordActivityType.cs
@@ -1,0 +1,30 @@
+namespace Discore
+{
+    public enum DiscordActivityType
+    {
+        /// <summary>
+        /// Example display: "Playing Overwatch".
+        /// </summary>
+        Game,
+        /// <summary>
+        /// Example display: "Streaming Overwatch".
+        /// </summary>
+        Streaming,
+        /// <summary>
+        /// Example display: "Listening to Spotify".
+        /// </summary>
+        Listening,
+        /// <summary>
+        /// Example display: "Watching YouTube Together".
+        /// </summary>
+        Watching,
+        /// <summary>
+        /// Example display: ":smiley: I am cool".
+        /// </summary>
+        Custom,
+        /// <summary>
+        /// Example display: "Competing in Arena World Champions".
+        /// </summary>
+        Competing,
+    }
+}